dashboard backend done

+++ dashboard-service/src/main/java/com/aeris/dashboard_service/dto/AlertDto.java
@@ -1,5 +1,5 @@
 // AlertDto.java
 package com.aeris.dashboard_service.dto;
 
-public record AlertDto(Integer id, String zone, String particule, String notation,
+public record AlertDto(Long id, String zone, String particule, String notation,
                         double concentration, String level, String timestamp) {}
